Almost done

+++ libft/ft_bzero.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 	char	*str;
 
 	i = 0;
