+++ libft/ft_lstdel.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	ft_lstdel(t_list **alst, void (*del)(void *, int))
+void	ft_lstdel(t_list **alst, void (*del)(void *, size_t))
 {
 	if (!(*alst))
 		return ;
