+++ libft/ft_lstnew.c
@@ -6,14 +6,14 @@
 /*   By: tshata <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/06 17:11:39 by tshata            #+#    #+#             */
-/*   Updated: 2018/07/30 17:28:54 by tshata           ###   ########.fr       */
+/*   Updated: 2018/06/23 13:02:11 by tshata           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-t_list	*ft_lstnew(void const *content, int content_size)
+t_list	*ft_lstnew(void const *content, size_t content_size)
 {
 	t_list *newlink;
 
@@ -23,7 +23,7 @@ t_list	*ft_lstnew(void const *content, int content_size)
 	newlink->next = NULL;
 	if (content != NULL)
 	{
-		newlink->content = (int *)malloc(content_size);
+		newlink->content = (size_t *)malloc(content_size);
 		if ((newlink->content) == NULL)
 			return (NULL);
 		ft_memcpy(newlink->content, content, content_size);
