+++ libft/ft_memchr.c
@@ -12,11 +12,11 @@
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*src;
 	unsigned char	ch;
-	int			i;
+	size_t			i;
 
 	src = (unsigned char *)s;
 	ch = (unsigned char)c;
