+++ libft/ft_strnstr.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, int n)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	if (haystack[i] == '\0' && needle[i] == '\0')
